Allow selecting which auth provider to use
This is a bit rubbish because you have to wait for the rel="me" stuff to be
worked out before seeing any response from the server.

I also added a `--true` flag which switches to using a always happy (local) auth
provider, it will not allow any other provider at the same time to prevent
accidents.